backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
requests==2.32.3
chardet<6
httpx==0.27.2
openai==1.55.3
google-generativeai==0.8.3
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
faster-whisper==1.1.0
SpeechRecognition==3.10.4
pyttsx3==2.98
python-dateutil==2.9.0.post0
cryptography==43.0.3
beautifulsoup4==4.12.3
reportlab==4.2.5
python-docx==1.1.2
matplotlib==3.9.2
plyer==2.1.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==4.1.0
